tests(smokehouse): Refactor so that Smokerider can run smoke tests.
